backend/requirements.txt
fastapi
python-dotenv
pydantic
fpdf
passlib[bcrypt]
nltk
openai
requests
snowflake-connector-python
pinecone-client
google-api-python-client
youtube-transcript-api
tavily-python
uvicorn
snowflake